fix: V9 dtype mismatch - cast bow_logits to float32 in loss computation
Training crashed with: mat1 and mat2 must have same dtype (Float vs BFloat16)

Root cause: bridge.bfloat16() converts bow_head weights to bfloat16,
but pooled.float() in forward created float32 input.

Fix:
- Remove .float() in forward, let bow_head stay in input dtype
- Cast bow_logits to float32 in training script before BCE loss

+++ telepathy/latent_bridge_v9.py
@@ -119,6 +119,6 @@ class LatentBridgeV9(nn.Module):
         # MAX pooling detects "Did this feature appear anywhere?"
         # Better than mean pooling for sparse entity detection
         pooled = compressed.max(dim=1).values  # [B, D]
-        bow_logits = self.bow_head(pooled.float())  # [B, Vocab]
+        bow_logits = self.bow_head(pooled)  # [B, Vocab] - stay in input dtype
 
         return scaled_out, bow_logits
